+++ core/src/main/kotlin/no/elg/hex/screens/PlayableIslandScreen.kt
@@ -313,7 +313,7 @@ class PlayableIslandScreen(metadata: FastIslandMetadata, island: Island) : Previ
     // we only ask for AI surrender when there is a single player
     // if there are no players (ai vs ai) we want to watch the whole thing
     // and if there are more than one player they should decide if they surrender
-    if (Settings.allowAIToSurrender && allowAISurrender && island.realPlayers == 1) {
+    if (Settings.allowAIToSurrender && allowAISurrender && island.singleAliveRealPlayer) {
       // surrender rules
       // either you own more than 75% of all hexagons
       // or all enemies have less than 12,5% of the hexagons